UxROM (mapper 2) and 8x16 sprites support

+++ src/Memory.cpp
@@ -3,15 +3,15 @@
 #include <cstdint>
 
 // TODO throw an error for an invalid address
-uint8_t Memory::read(uint16_t addr) {
+uint8_t Memory::read(uint32_t addr) {
     return addr < size ? mem[addr] : 0x0;
 }
 
-bool Memory::write(const uint16_t addr, const uint8_t data) {
+bool Memory::write(const uint32_t addr, const uint8_t data) {
     return this->write(addr, &data, 1);
 }
 
-bool Memory::write(const uint16_t addr, const uint8_t* data, const int bytes) {
+bool Memory::write(const uint32_t addr, const uint8_t* data, const int bytes) {
     if (addr + bytes <= size) {
         mem[addr] = *data;
         return true;
@@ -19,7 +19,7 @@ bool Memory::write(const uint16_t addr, const uint8_t* data, const int bytes) {
     return false;
 }
 
-void Memory::init(const uint16_t size) {
+void Memory::init(const uint32_t size) {
     delete[] this->mem;
     this->size = size;
     this->mem = new uint8_t[size];
